Add attempt parameter, and allow no arguments to be passed to the worker
