+++ src/error.ts
@@ -1,0 +1,7 @@
+// error.ts
+
+export class RetryError extends Error {
+  constructor(public retries: number, public lastError: Error) {
+    super(`Maximum retries (${retries}) exceeded`);
+  }
+}
