+++ src/options.ts
@@ -1,0 +1,7 @@
+// options.ts
+
+export interface RetryOptions {
+  waitRatio?: number;
+  retries?: number;
+  jitter?: boolean;
+}
